Weekly calories and today meals displayed. photo size reduced

+++ app/src/main/java/com/sks225/snapeat/model/Common.kt
@@ -5,7 +5,7 @@ data class Common(
     val food_name: String,
     val locale: String,
     val photo: Photo,
-    val serving_qty: Int,
+    val serving_qty: Double,
     val serving_unit: String,
     val tag_id: String,
     val tag_name: String
